moved the activity thingies inside of activities folder
