+++ app/src/androidTest/java/com/sdp13epfl2021/projmag/ProfilePageActivityTest.kt
@@ -9,6 +9,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
 import androidx.test.runner.AndroidJUnit4
+import com.sdp13epfl2021.projmag.activities.ProfilePageActivity
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.`is`
