+++ app/src/androidTest/java/com/sdp13epfl2021/projmag/ProfilePageTestEdit.kt
@@ -4,6 +4,7 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.sdp13epfl2021.projmag.activities.ProfilePageActivity
 import org.junit.Rule
 import org.junit.Test
