+++ app/src/androidTest/java/com/sdp13epfl2021/projmag/SignInActivityTest.kt
@@ -6,6 +6,7 @@ import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.runner.AndroidJUnit4
+import com.sdp13epfl2021.projmag.activities.SignInActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
