+++ app/src/main/AndroidManifest.xml
@@ -14,14 +14,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Projmag">
         <activity android:name=".activities.tagsSelectorActivity"></activity>
-        <activity android:name=".ProfilePageActivity" />
+        <activity android:name=".activities.ProfilePageActivity" />
         <activity
             android:name=".curriculumvitae.CVCreationActivity"
             android:screenOrientation="portrait" />
         <activity
             android:name=".activities.ProjectInformationActivity"
             android:parentActivityName=".activities.ProjectsListActivity" />
-        <activity android:name=".SignInActivity" />
+        <activity android:name=".activities.SignInActivity" />
         <activity android:name=".Form" />
         <activity android:name=".activities.ProjectsListActivity" />
         <activity android:name=".MainActivity">
