+++ app/src/main/java/com/sdp13epfl2021/projmag/activities/SignInActivity.kt
@@ -1,4 +1,4 @@
-package com.sdp13epfl2021.projmag
+package com.sdp13epfl2021.projmag.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -12,7 +12,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.gms.common.api.ApiException
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
-import com.sdp13epfl2021.projmag.activities.ProjectsListActivity
+import com.sdp13epfl2021.projmag.R
 
 
 /**
